Added new buttons to save background pc and get board dimensions

+++ include/cam_lidar_calibration/point_xyzir.h
@@ -9,12 +9,12 @@ struct PointXYZIR
 {
   PCL_ADD_POINT4D;
   float intensity;
-  uint16_t ring;
+  uint8_t ring;
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 } EIGEN_ALIGN16;
 }  // namespace pcl
 
 POINT_CLOUD_REGISTER_POINT_STRUCT(PointXYZIR, (float, x, x)(float, y, y)(float, z, z)(float, intensity,
-                                                                                      intensity)(uint16_t, ring, ring))
+                                                                                      intensity)(uint8_t, ring, ring))
 
 #endif
